Renaming contrib to owpcontrib.
Modified the OWP_CONF_DIR flag to be set relative to a "prefix" that is set
as autoconf normally would. - this took renaming access.h to access.h.in and
having autoconf parse it and make changes.

api.c:
Added a connect_tmout to api.c

Corrected logic for CheckControlPolicy checks.

owping:
A bit more - won't currently compile.

jeff


git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@144 a2d82226-602c-0410-bba7-e51b83dd10d2