Big changes... Too many to go through. I should not have gone so long between check-in's.
Basically I think the control protocol is nearly finished. I have not actually
tried it yet however - it is actually a little hard to test without firing off
a test, and I have not written the udp code yet. tomorrow. The only part of
the control protocol that is not finished is retreiving session data. The
client can currently request the data, but the server will always deny.
(I figured we can just look at the data on disk for now - I want to get
actual sessions working before messing with it.)

removed the CheckAddr func stuff.

jeff


git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@241 a2d82226-602c-0410-bba7-e51b83dd10d2