Modified the app_data to be relative to a connection. It should be set with the data type defiend in owamp/conndata.h - which I will check in momentarily.
This will allow each child process to request resouces from the parent process.
(It might be a cool idea to call the policy stuff directly if pipefd is < 0.
that way the same functions can be used from the client but we can talk
about this when I get back.)

Tolya, I hope this is clear to you.


See you next week. I need to take off now.

jeff


git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@244 a2d82226-602c-0410-bba7-e51b83dd10d2