Added a skeleton usage() - will rework it further, though.
Added function
/*
** Fetch a session with the given <sid> from the remote server.
** It is assumed that control connection has been opened already.
*/
void
owp_fetch_sid(
	      char *savefile,
	      OWPControl cntrl,
	      OWPSID sid,
	      fetch_state_ptr statep,
	      char *local,
	      char *remote
	      )

- currently used by regular owping (when client == sender)
but can be easily used by  owfetch. Further refactoring
is possible and may well happen in fact.

Currently it's not very clear how to handle the fetch
of multiple SIDs within a single control session
(since we have to allow for them to be saved I don't see
how the command-line can even be parsed in this case).
If this can be arranged then it's enough to simply
call owp_fetch_sid() multiple times.


git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@392 a2d82226-602c-0410-bba7-e51b83dd10d2