Modified to use high 3 order bytes of frac_sec(32) instead of low order bytes. Hopefully things are consistant now.
(Inconsistant use of this was causing packets to be sent before the start
time.)

also fixed a race condition in endpoint that happened if the endpoint proc
happen to exit after its status was queried, but before it was killed.

modified owping/owstats -V printout to use OWP_TSTAMPFMT.

(temporarily removed file-locking from powstream because restarts in
gdb don't free the locks. VERY ANNOYING!)

jeff


git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@435 a2d82226-602c-0410-bba7-e51b83dd10d2