Unified "time" api code so all time functions go through the functions in time.c.
Created version 3 of the owd file format to support ttl data in the records.

Modified the code to set the ttl and retrieve the ttl.

Created single functions for encoding and decoding data records to make it
more simple to change the file i/o functions.

Simplified the "receiver" code by using file i/o functions.

Modified the way "lost packet" records are encoded in the protocol data
stream and in files.

removed functions from unixtime.c

(all this still needs to be compiled and tested...)

jeff


git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@769 a2d82226-602c-0410-bba7-e51b83dd10d2