Bug fixes for new code:
Needed to modify where the datafile is closed. (It used to be closed
in the parent immediately, now the parent has to go through the data
when the stop sessions message comes in to delete records after
(stoptime - timeout).

Also needed to modify all fdopens to open w+ instead of w so the
records can be queried as well as written.

Needed to modify the interpolated binary search to round instead of truncate
when determining the candidate.

Missed a few spots where the return type of the DataHeader write functions
changed return values. (They now return boolean - they used to return int's.)

Receive sessions are working from owping now. There is still a problem with
the Fetch protocol at the moment that I have not tracked down yet.

jeff


git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@805 a2d82226-602c-0410-bba7-e51b83dd10d2