Corrected FetchSession interaction. Needed to send block of IZP between skips and data records even if there are no skips. I had accidentily put the write of that block in the if(num_skiprecs) block.
Simplified the Data file reading code slightly.

Corrected FetchSession interaction. Needed to send block of IZP between
skips and data records even if there are no skips. I had accidentily put the
write of that block in the if(num_skiprecs) block.

Simplified the Data file reading code slightly.

Still to be done for this version (aiming for this week):

        Fix shared memory IPC - just need to add an fdtruncate call
        to assign the file size. (Currently using normal files for
        IPC - this is a compile time selection.)

        Add ttl summary info to owping/owstats output.

        Test ttl data on an OS that supports it...

        Add sub-session summaries to powstream so AMI database host
        interaction can be optimized.

        Porting to additional OS's.

jeff


git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@806 a2d82226-602c-0410-bba7-e51b83dd10d2