General: Modified code to remove compiler warnings. reformatted some of the code to remove annoying tabs. (Will go through and reformat the rest before releasing.)
powstream.c:
    Major rework - simplified the code (somewhat). Summary files
    are now the "intermediate" data output by powstream. It only
    puts out raw data files based on the larger interval you give
    it. (On abilene that is currently 12 hours... we may want to
    change that.)

    Summary statistics will be put out more regularly. I suspect
    we will stick with the 30 second summary sessions for abilene.


**** Not TESTED! ****

It compiles completely clean, but I have not had a chance to run it
yet. (It does compile on OS X as well... I guess this is what happens
when that is what is sitting in front of me.)

owamp plans:

Testing starts next week along with major reworking of the ami
perl scripts to handle the ttl data as well as the new interval
setup of powstream. (I intend to gut most of the "validation" code
from the ami scripts and use the "skip records" from the raw data
files for this. It should make the database code about 50% easier!)

jeff


git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@831 a2d82226-602c-0410-bba7-e51b83dd10d2