Adding two new -M output fields:
SESSION_PACKET_COUNT
SAMPLE_PACKET_COUNT

It turns out it is useful to put some amount of the parameters from the
test request in the output data. Especially since it is being sampled,
by owstats... I may need to add more of these to fully generalize
the powcollector implementation, but this is enough for a first shot.


git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@1064 a2d82226-602c-0410-bba7-e51b83dd10d2