Adding nearly all test params as output to the -M format. (Don't do the schedule, because that *could* be complex and I'm not sure it is a good idea. pS-B will only support a single 'interval' value for now anyway.)
Also removed references to the outdated -H parameter to owping.

In my testing, I was shocked that IPV6_TCLASS Diff-serv bits are still not
supported in MacOS. So I looked around, and even though this made RFC
status in 2003, FreeBSD didn't even get this until last year - and Linux
was only the year before that. So, if we are going to try using
scavenger service for testing, we will need to carefully check if
the OSs support it or not.



git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@1066 a2d82226-602c-0410-bba7-e51b83dd10d2