Pull the changes from the powstream_loopissue branch into trunk:
r1080:

Bump the -rc version # 

r1078:

Switch to a per-cntrl fetch option, and reset the session in SetupSession. I'm
not positive this doesn't screw with the semantics, but I'm committing so
others can read it while I'm testing to make sure I'm not breaking something.

r1076:

On error, powstream kept opening new fetch connections, but not closing the old fds.

r1074:

Change powstream to have it pause before between retrying tests. This is just a first cut.




git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@1081 a2d82226-602c-0410-bba7-e51b83dd10d2