We saw some issues in the field where powstream was displaying that it was going to sleep for negative time durations, and then would try to do so using nanosleep.
There was code to only print out the sleep message if the duration was > 0,
but, on some hosts, the comparison time_t of with 0 claim the time value was >
0, no matter what the actual value was. To make matters worse, when the
timespec containing the negative seconds value was passed to nanosleep, it
would sleep for an indeterminate amount of time. My guess is that it treated 
the seconds value as unsigned and was sleep for a ridiculous number of seconds.

This patch changes how the nanosleep time gets calculated. The check for the
next runtime is calculated using the OWPNum time values instead of the timespec
values. Also, nanosleep is only called if the next runtime is later than now.



git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@1083 a2d82226-602c-0410-bba7-e51b83dd10d2