Client TWAMP control protocol handling
Adapt twping so that it requests just one two-way test session, rather
than a sender and a receiver session that owping does. Remove command
line options that don't make sense as a result.

Don't send a SID in the session request message for TWAMP. Ensure it's
filled in based on the value in the response.

Disallow use of OWPFetchSession for TWAMP control sessions.