Add directional delay statistics for TWAMP
The min, median and max delay values for the forward and backward
directions are now displayed after a twping session.

PDV for the forward and backward directions is also now shown.

Min/max processing delay is now shown in the machine readable output.

Median values for all directions are now displayed in the machine
readable output.

PDV is now shown in the machine readable output.