Remove unncessary getnameinfo() calls
Previously in OWPStatsCreate(), getnameinfo() was always called twice
to retrieve a hostname for the sender and receiver. However, if
fromhost or tohost were passed in (and weren't NULL) then the
getnameinfo() result would always be overwritten with those values,
making the getnameinfo() calls unnecessary.

For twping fromhost and tohost should usually be set, so we now only
call getnameinfo() to retrieve the hostname when the corresponding
value isn't set.