Set Receiver Port field in Request-TW-Session msg
Previously the Receiver Port field was set to all zeros which some
reflectors do not support. Such reflectors accepted the zero port
without offering an alternative and caused the test session to fail.

Now we request that the reflector use the same test port number that
was chosen for the local test port. If this is not acceptable to the
reflector it may offer an alternative which we will then use.