Make use of IPv6 link-local addresses for test session endpoints work.
Prior to this, when a client requested an IPv6 link-local address to
be used for a test session endpoint, the bind to the address would
fail since a link-local address isn't valid without a scope indicating
which link (interface) it applies to.

Therefore, to support test sessions listening on IPv6 link-local
addresses, the scope needs to be specified. The only way to get this
is from the control session remote endpoint's sockaddr, so copy the
scope over if available or otherwise the only option is to reject the
request.