Reject session if the test port cannot be determined
If the port number cannot be retrieved then instead of sending an
Accept-Sessions message with a 0 Port and 0 Accept, we whould reject
the test session by sending an Accept-Session message with an Accept
value of 2 (indicating internal failure/error).

As per RFC 5357 Section 3.5, the Port field must be zero when Accept
is non-zero. This change is consistent with this constraint.