twamp: Send negative accept for all RequestTWSessions errors
RFC 5357 s3.5 specifies that a negative accept message MUST be sent if
either the Conf-Sender or Conf-Receiver fields are non-zero. The
checking for this is done in _OWPReadTestRequest, so if this returns a
warning then this should be translated into sending the negative
accept message instead of just returning to the caller.