Add dynamic capabilities support using libcap
This is used to dynamically raise the CAP_NET_RAW capability into
the effective set in order to use SO_BINDTODEVICE.

This allows the -B option to [ot]wping and powstream to be used by
users without the CAP_NET_RAW privilege, typically non-root users.

To make use of the functionality you would typically set CAP_NET_RAW
in a file's permitted capability set eg.
  $ setcap "cap_net_raw+p" /usr/bin/twping

This functionality is only enabled when libcap is present in the
build environment.