API level support for SO_BINDTODEVICE
Adds new APIs for binding sockets to network interfaces.

OWPSocketInterfaceBind() sets SO_BINDTODEVICE with a given interface
name on a given socket.

OWPControlOpenInterface() and TWPControlOpenInterface() are identical
to OWPControlOpen() and TWPControlOpen() except for an additional
interface argument which specifies the interface to bind to. The
OWAMP/TWAMP control connection created by the new APIs is bound to
the specified interface. Sockets for test sessions under those
control sessions will also be bound to the same interface.