twping: Send zero-addresses in request to support NAT
This commit allows a client behind a NAT and the server behind a NAT.
The server's NAT must forward the TCP control port and UDP test ports to
the server.

When the twping client request message contains zero-addresses, the
twampd server will use the addresses from the TCP control connection.
This is specified in RFC 5357 section 3.5. From the perspective of the
server the endpoints are the server's internal IP address and the
client's public (NAT) IP address. This allows the server to send the
packets back to the correct address. The client could not provide these
addresses in the request message.