Fix test/owstats for 32bit architecture (i386)
owamp/config.h does a #define _FILE_OFFSET_BITS 64, which changes the
size of off_t from 4 to 8 in i386 architectures, and also the size of
struct stat. That, in turn, changes the size of the structures in
owamp/owamp.h that have off_t and struct stat members.

test/owstats was including fcntl.h and others before including
owamp/config.h, and that made it create shorter structures in the stack.
When the test called api.c functions that had the structure with the
correct size, the stack was overrun, generating a segmentation fault.

This commit adds a #include <owamp/owamp.h> at the start of all tests
to avoid this kind of problem. The other C files in this project already
do that, and the files that come from I2util don't define structs with
these members.